Add example submission XML and stub script
